requirements.txt
Django==3.0.6
PyMySQL==0.9.3
cryptography==2.9.2
django-cors-headers==3.2.1
djangorestframework==3.11.0